Added levels, gliding, and gravity
Added main level that features enemies and two feathers, one that adds a second jump and one that gives glide time. Completing this level leads the player to an end screen.

Gliding is functional, and keeps the player's vertical speed at a small static number while holding down W.

Look into bugs involving falling state set to false when hitting tiles from the side or below (enabling jumping and setting vspeed to 0), and enemies not being able to check if a tile is in front.