Added dialogue system
Added dialogue system, which currently only displays when collecting a feather.

Can display up to three lines of dialogue (wouldn't be hard to add more if we need it though). When creating dialogue, it must be defined how many lines it will have and what those lines display.

Once created, it will display the text in a box on the top of the screen one letter at a time. The player can press enter to skip this and make it display the full line.

Once the line is fully displayed, the player can press enter to either move to the next line or close the dialogue box.