Added downed enemies and enemy eating
After depleting an enemy's health, it will be downed, flipped upside down. After a variable number of seconds (depending on the enemy) or if you hit it again, it ill be destroyed.

If you press H next to the enemy (to be restricted later), it will be 'eaten' and you will recover a small amount of health.