Added tutorial area
Tutorial area introduces platforming, attacking with sword and birdsong, and how to catch worms (to be implemented).

NPCs are present, each of whom introduce a concept to the player by interacting with them by pressing Enter.

The final NPC warps the player to the start of the game.

Also added room border objects, which are used to move the player between rooms.