Added worm, extra NPCs, and cheats
Added worm object that will attempt to run away from the player. Catching it will restore health, restoring extra if player health is low.

Added two NPCs in the main area, one explaining that the player's wings are damaged and need to collect feathers, and the other exposing the top passageway.

Added cheats: 0 = go to tutorial area, 1 = go to first area, Tab = high speed and lots of jumps.

Also added a passageway that will lead to another section in the main area.